Fix: some clang tidy warnings

+++ src/lib/covercache.c
@@ -66,7 +66,7 @@ bool covercache_write_file(const char *workdir, const char *uri, const char *mim
 int covercache_clear(const char *workdir, int keepdays) {
     int num_deleted = 0;
     bool error = false;
-    time_t expire_time = time(NULL) - keepdays * 24 * 60 * 60;
+    time_t expire_time = time(NULL) - (long)(keepdays * 24 * 60 * 60);
     
     sds covercache = sdscatfmt(sdsempty(), "%s/covercache", workdir);
     MYMPD_LOG_NOTICE("Cleaning covercache %s", covercache);
